+++ src/main/java/ch/zhaw/dna/ssh/mapreduce/model/WordFrequencyCombinerInstruction.java
@@ -28,7 +28,7 @@ public class WordFrequencyCombinerInstruction implements CombinerInstruction {
 				KeyValuePair formerKeyValuePair = combinedKeyValuePairs.get(currentKeyValuePair.getKey());
 				combinedKeyValuePairs.remove(currentKeyValuePair.getKey());
 				
-				combinedKeyValuePairs.put(currentKeyValuePair.getKey(), new KeyValuePair(currentKeyValuePair.getKey(), "" + 1 + Integer.parseInt(formerKeyValuePair.getValue())));
+				combinedKeyValuePairs.put(currentKeyValuePair.getKey(), new KeyValuePair(currentKeyValuePair.getKey(), "" + 1 + Long.parseLong(formerKeyValuePair.getValue())));
 				
 			} else {
 				combinedKeyValuePairs.put(currentKeyValuePair.getKey(), currentKeyValuePair);
